Add very short variant display names to comparison table
