+++ apps/store/public/locales/en/purchase-form.json
@@ -11,6 +11,9 @@
   "CHANGE_SSN_DESCRIPTION": "Products you have already added to your shopping cart will not be saved.  Form data will not be saved. You may need to login with BankID again.",
   "CHANGE_SSN_TITLE": "Change personal number?",
   "COMPARE_COVERAGE_BUTTON": "Compare coverage",
+  "COMPARISON_TABLE_COLUMN_SE_CAR_FULL": "Full",
+  "COMPARISON_TABLE_COLUMN_SE_CAR_HALF": "Half",
+  "COMPARISON_TABLE_COLUMN_SE_CAR_TRAFFIC": "Traffic",
   "CURRENT_INSURANCE_FIELD_PLACEHOLDER": "Pick company...",
   "FIELD_ANCILLARY_AREA_LABEL": "Ancillary area",
   "FIELD_CAR_REGISTRATION_NUMBER_LABEL": "Registration number",
