+++ apps/store/public/locales/sv-se/purchase-form.json
@@ -11,6 +11,9 @@
   "CHANGE_SSN_DESCRIPTION": "Produkter som du redan har lagt till i din varukorg kommer inte att sparas. Formulärdata kommer inte att sparas. Du kan behöva logga in med BankID igen.",
   "CHANGE_SSN_TITLE": "Ändra personnummer?",
   "COMPARE_COVERAGE_BUTTON": "Jämför skydden",
+  "COMPARISON_TABLE_COLUMN_SE_CAR_FULL": "Hel",
+  "COMPARISON_TABLE_COLUMN_SE_CAR_HALF": "Halv",
+  "COMPARISON_TABLE_COLUMN_SE_CAR_TRAFFIC": "Trafik",
   "CURRENT_INSURANCE_FIELD_PLACEHOLDER": "Välj bolag...",
   "FIELD_ANCILLARY_AREA_LABEL": "Biyta",
   "FIELD_CAR_REGISTRATION_NUMBER_LABEL": "Registreringsnummer",
